+++ readthedocs/filetreediff/dataclasses.py
@@ -1,5 +1,12 @@
 from dataclasses import asdict
 from dataclasses import dataclass
+from enum import StrEnum
+from enum import auto
+from functools import cached_property
+
+from readthedocs.builds.models import Build
+from readthedocs.builds.models import Version
+from readthedocs.core.resolver import Resolver
 
 
 @dataclass(slots=True)
@@ -10,7 +17,7 @@ class FileTreeDiffBuild:
 
 
 @dataclass(slots=True)
-class FileTreeDiffFile:
+class FileTreeDiffManifestFile:
     """A file in a file tree manifest."""
 
     path: str
@@ -21,10 +28,10 @@ class FileTreeDiffFile:
 class FileTreeDiffManifest:
     """A list of files and the build associated with them."""
 
-    files: dict[str, FileTreeDiffFile]
+    files: dict[str, FileTreeDiffManifestFile]
     build: FileTreeDiffBuild
 
-    def __init__(self, build_id: int, files: list[FileTreeDiffFile]):
+    def __init__(self, build_id: int, files: list[FileTreeDiffManifestFile]):
         self.build = FileTreeDiffBuild(id=build_id)
         self.files = {file.path: file for file in files}
 
@@ -38,7 +45,7 @@ class FileTreeDiffManifest:
         """
         build_id = data["build"]["id"]
         files = [
-            FileTreeDiffFile(path=path, main_content_hash=file["main_content_hash"])
+            FileTreeDiffManifestFile(path=path, main_content_hash=file["main_content_hash"])
             for path, file in data["files"].items()
         ]
         return cls(build_id, files)
@@ -48,14 +55,105 @@ class FileTreeDiffManifest:
         return asdict(self)
 
 
-@dataclass
+class FileTreeDiffFileStatus(StrEnum):
+    """Status of a file in the file tree diff."""
+
+    ADDED = auto()
+    DELETED = auto()
+    MODIFIED = auto()
+
+    @property
+    def emoji(self) -> str:
+        """Return an emoji representing the file status."""
+        if self == FileTreeDiffFileStatus.ADDED:
+            return "➕"
+        elif self == FileTreeDiffFileStatus.DELETED:
+            return "❌"
+        elif self == FileTreeDiffFileStatus.MODIFIED:
+            return "📝"
+        return ""
+
+
+class FileTreeDiffFile:
+    def __init__(
+        self,
+        path: str,
+        status: FileTreeDiffFileStatus,
+        diff: "FileTreeDiff",
+    ):
+        self.path = path
+        self.status = status
+        self.current_version = diff.current_version
+        self.current_version_build = diff.current_version_build
+        self.base_version = diff.base_version
+        self.base_version_build = diff.base_version_build
+        self._resolver = diff._resolver
+
+    @property
+    def url(self):
+        """URL to the file in the current version."""
+        return self._resolver.resolve_version(
+            project=self.current_version.project,
+            version=self.current_version,
+            filename=self.path,
+        )
+
+    @property
+    def base_url(self):
+        """URL to the file in the base version."""
+        return self._resolver.resolve_version(
+            project=self.base_version.project,
+            version=self.base_version,
+            filename=self.path,
+        )
+
+
 class FileTreeDiff:
-    """Difference between two file tree manifests."""
+    def __init__(
+        self,
+        current_version: Version,
+        current_version_build: Build,
+        base_version: Version,
+        base_version_build: Build,
+        files: list[tuple[str, FileTreeDiffFileStatus]],
+        outdated: bool,
+    ):
+        self.current_version = current_version
+        self.current_version_build = current_version_build
+        self.base_version = base_version
+        self.base_version_build = base_version_build
+        self._resolver = Resolver()
+        self.outdated = outdated
+        self.files = sorted(
+            (FileTreeDiffFile(path=path, status=status, diff=self) for path, status in files),
+            key=self._sortpath,
+        )
 
-    added: list[str]
-    deleted: list[str]
-    modified: list[str]
-    outdated: bool = False
+    def _sortpath(self, file: FileTreeDiffFile):
+        """
+        Function to use as `key=` argument for `sorted`.
 
-    def has_changes(self) -> bool:
-        return bool(self.added or self.deleted or self.modified)
+        It sorts the file names by the less deep directories first.
+        However, it doesn't group the results by directory.
+
+        Ideally, this should sort file names by hierarchy (less deep directory
+        first), groupping them by directory and alphabetically. We should update
+        this function to achieve that goal if we find a simple way to do it.
+        """
+        parts = file.path.split("/")
+        return len(parts), parts
+
+    @cached_property
+    def added(self):
+        """List of added files."""
+        return [file for file in self.files if file.status == FileTreeDiffFileStatus.ADDED]
+
+    @cached_property
+    def deleted(self):
+        """List of deleted files."""
+        return [file for file in self.files if file.status == FileTreeDiffFileStatus.DELETED]
+
+    @cached_property
+    def modified(self):
+        """List of modified files."""
+        return [file for file in self.files if file.status == FileTreeDiffFileStatus.MODIFIED]
